[Editorial] Make "Poor Knight" wordplay consistent with Garnett
There is some wordplay in part 2, chapter 6 and 7 where Aglaya changes
Pushkin's poem "The Poor Knight" to replace the knight's abbreviated
motto with "N. P. B." (the initials of another character in the book).
Martin's translation seems to be flawed here because she uses "A.N.D."
and "A.N.B." in two places where Garnett uses "A.M.D." Garnett's usage
makes much more sense because my research suggests that in the original
poem, the motto on the knight's shield is supposed to be "Ave Mater
Dei." For cross-referencing, consult pages 248-251 of Garnett's
translation (https://archive.org/details/in.ernet.dli.2015.549420) and
pages 236-238 of Martin's translation.